requirements.txt
boto3==1.14.40
botocore==1.17.40
click==7.1.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
